[#683] Merge Writer Dashboard into Profile Stories tab
- Profile Stories tab now shows full writer dashboard content (genre editing,
  claim royalties, trading stats, deadline countdown, donation history) when
  viewing own profile
- Other users' profiles show storylines read-only with token price + TVL
- Fix BigInt(0) rendering bug in wallet card (claimedRoyalties != null check)
- /writer page now redirects to /profile/[address]?tab=stories
- Profile page reads ?tab= query param for deep linking
